requirements_transfer.txt
absl-py==2.3.1
ale-py==0.9.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
astunparse==1.6.3
AutoROM==0.6.1
AutoROM.accept-rom-license==0.6.1
av==16.0.1
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
chex==0.1.90
click==8.3.0
cloudpickle==3.1.2
colored-traceback==0.4.2
contourpy==1.3.3
control-msgs==4.8.0
controller-manager==2.51.0
controller-manager-msgs==2.51.0
crafter==1.8.3
cv-bridge==3.2.1
cycler==0.12.1
decorator==5.2.1
demo-nodes-py==0.20.5
diagnostic-msgs==4.9.0
diagnostic-updater==4.0.6
domain-coordinator==0.10.0
-e git+ssh://git@github.com/xavier2933/dreamerv3.git@c89b66e5db87dabfafabca788a7207020fe9a201#egg=dreamer
einops==0.8.1
elements==3.20.0
# Editable install with no version control (embodied==2.0.0)
executing==2.2.1
fastapi==0.121.0
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.6.0
gazebo-msgs==3.9.0
generate-parameter-library-py==0.5.0
geographic-msgs==1.0.6
geometry-msgs==4.9.0
google-api-core==2.28.1
google-auth==2.42.1
google-cloud-core==2.5.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
granular==0.21.0
graph-msgs==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
httptools==0.7.1
idna==3.11
image-geometry==3.2.1
ImageIO==2.37.2
interactive-markers==2.3.2
ipdb==0.13.13
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jax==0.4.33
jax-cuda12-pjrt==0.4.33
jax-cuda12-plugin==0.4.33
jaxlib==0.4.33
jaxtyping==0.3.3
jedi==0.19.2
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
keras==3.12.0
kiwisolver==1.4.9
laser-geometry==2.4.0
launch==1.0.10
launch-param-builder==0.1.1
launch-testing==1.0.10
launch-xml==1.0.10
launch-yaml==1.0.10
libclang==18.1.1
lifecycle-msgs==1.2.2
logging-demo==0.20.5
map-msgs==2.1.0
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mediapy==1.2.4
message-filters==4.3.8
ml_dtypes==0.5.3
moveit-configs-utils==2.5.9
moveit-msgs==2.2.1
msgpack==1.1.2
namex==0.1.0
nav-msgs==4.9.0
ninjax==3.6.2
numpy==1.26.4
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvcc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.15.0.57
nvidia-cufft-cu12==11.4.1.4
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-nccl-cu12==2.28.7
nvidia-nvjitlink-cu12==12.9.86
object-recognition-msgs==2.0.0
octomap-msgs==2.0.1
opensimplex==0.4.5.1
opt_einsum==3.4.0
optax==0.2.5
optree==0.18.0
orjson==3.11.4
osrf-pycommon==2.1.6
packaging==25.0
pandas==2.3.3
parso==0.8.5
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pexpect==4.9.0
pillow==12.0.0
portal==3.7.3
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
py-trees==2.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-qt-binding==1.1.2
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.5
rcutils==5.1.6
requests==2.32.5
resource-retriever==3.1.3
rich==14.2.0
rmw-dds-common==1.6.0
robot-localization==3.5.3
rpyutils==0.2.1
rqt==1.1.7
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-controller-manager==2.51.0
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-joint-trajectory-controller==2.49.1
rqt-moveit==1.0.1
rqt-msg==1.2.0
rqt-plot==1.1.5
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==1.0.6
rqt-robot-steering==1.0.3
rqt-runtime-monitor==1.0.0
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-tf-tree==1.0.5
rqt-topic==1.5.0
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
scipy==1.16.3
scope==0.6.3
scripts==3.9.0
sensor-msgs==4.9.0
sensor-msgs-py==4.9.0
shape-msgs==4.9.0
six==1.17.0
sniffio==1.3.1
srdfdom==2.0.7
sros2==0.10.7
stack-data==0.6.3
starlette==0.49.3
statistics-msgs==1.2.2
std-msgs==4.9.0
std-srvs==4.9.0
stereo-msgs==4.9.0
teleop-twist-keyboard==2.4.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
tensorflow_cpu==2.20.0
termcolor==3.2.0
tf-transformations==1.1.0
tf2-geometry-msgs==0.25.16
tf2-kdl==0.25.16
tf2-msgs==0.25.16
tf2-py==0.25.16
tf2-ros-py==0.25.16
tf2-sensor-msgs==0.25.16
tf2-tools==0.25.16
toolz==1.1.0
topic-monitor==0.20.5
tqdm==4.67.1
traitlets==5.14.3
trajectory-msgs==4.9.0
turtlesim==1.4.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unique-identifier-msgs==2.2.1
urdfdom-py==1.2.1
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
visualization-msgs==4.9.0
wadler_lindig==0.1.7
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.3
wrapt==2.0.1
xacro==2.0.13
